Refactor to avoid globals. Always display on LCD.
This patch significantly changes the structure of the code. There is now
a main() function and the background thread is gone.

To facilitate the removal of global variables there is now an LCD class
which stores the lines to be displayed and ensures they are padded to 16
chars.

MQTT's own background thread is used to dispatch messages so that we
don't have to handle reconnecting to mqtt ourselves.

Since using ioctl to get the IP address of the interfaces is much
faster we can now poll every second.

Always display data on the LCD, even if the backlight is off.